UserService persistence unified & Working properly storing statistics
