+++ users/userservice/user-service.js
@@ -1,22 +1,37 @@
-// user-service.js
 const express = require('express');
 const mongoose = require('mongoose');
 const bcrypt = require('bcrypt');
-const User = require('./user-model')
+const cors = require('cors');
+const User = require('./user-model');
+const GameStats = require('./game-stats-model');
 
 const app = express();
-const port = 8001;
+const port = process.env.PORT || 8001;
 
-// Middleware to parse JSON in request body
 app.use(express.json());
 
-// Connect to MongoDB
+app.use(cors({
+  origin: process.env.CORS_ORIGIN || 'http://localhost:3000',
+  methods: ['GET', 'POST'],
+  allowedHeaders: ['Content-Type']
+}));
+
 const mongoUri = process.env.MONGODB_URI || 'mongodb://localhost:27017/userdb';
-mongoose.connect(mongoUri);
+mongoose.connect(mongoUri)
+  .then(async () => {
+    console.log('Conectado a MongoDB');
+    
+    // Comprobamos si la colección 'stats' existe, si no, la creamos.
+    const collections = await mongoose.connection.db.listCollections({ name: 'stats' }).toArray();
+    if (collections.length === 0) {
+      await mongoose.connection.db.createCollection('stats');
+      console.log("Colección 'stats' creada con éxito.");
+    } else {
+      console.log("Colección 'stats' ya existe.");
+    }
+  })
+  .catch(err => console.error('Error de conexión a MongoDB:', err.message));
 
-
-
-// Function to validate required fields in the request body
 function validateRequiredFields(req, requiredFields) {
     for (const field of requiredFields) {
       if (!(field in req.body)) {
@@ -25,12 +40,10 @@ function validateRequiredFields(req, requiredFields) {
     }
 }
 
+// Ruta para crear un usuario (se guarda en la colección 'users')
 app.post('/adduser', async (req, res) => {
     try {
-        // Check if required fields are present in the request body
         validateRequiredFields(req, ['username', 'password']);
-
-        // Encrypt the password before saving it
         const hashedPassword = await bcrypt.hash(req.body.password, 10);
 
         const newUser = new User({
@@ -42,16 +55,64 @@ app.post('/adduser', async (req, res) => {
         res.json(newUser);
     } catch (error) {
         res.status(400).json({ error: error.message }); 
-    }});
+    }
+});
+
+// Ruta para guardar estadísticas (se guarda en la colección 'stats')
+app.post('/api/stats', async (req, res) => {
+  try {
+    const { correctAnswers, incorrectAnswers, totalRounds } = req.body;
+    if (!correctAnswers || !incorrectAnswers || !totalRounds) {
+      return res.status(400).json({
+        error: 'Bad Request',
+        message: 'Faltan campos requeridos: correctAnswers, incorrectAnswers, totalRounds'
+      });
+    }
+
+    const newStats = new GameStats({
+      correctAnswers,
+      incorrectAnswers,
+      totalRounds,
+      accuracy: parseFloat(((correctAnswers / totalRounds) * 100).toFixed(2))
+    });
+
+    const savedStats = await newStats.save();
+    res.status(201).json(savedStats);
+
+  } catch (error) {
+    console.error('Error guardando estadísticas:', error);
+    res.status(500).json({
+      error: 'Internal Server Error',
+      message: error.message
+    });
+  }
+});
+
+// Ruta para obtener estadísticas (se obtienen de la colección 'stats')
+app.get('/api/stats', async (req, res) => {
+  try {
+    const limit = parseInt(req.query.limit) || 10;
+    const stats = await GameStats.find()
+      .sort({ timestamp: -1 })
+      .limit(limit);
+
+    res.json(stats);
+
+  } catch (error) {
+    console.error('Error obteniendo estadísticas:', error);
+    res.status(500).json({
+      error: 'Internal Server Error',
+      message: error.message
+    });
+  }
+});
 
 const server = app.listen(port, () => {
   console.log(`User Service listening at http://localhost:${port}`);
 });
 
-// Listen for the 'close' event on the Express.js server
 server.on('close', () => {
-    // Close the Mongoose connection
-    mongoose.connection.close();
-  });
+  mongoose.connection.close();
+});
 
-module.exports = server
+module.exports = server;
