+++ webapp/src/components/wikidataComponents/QuestionPresentation.jsx
@@ -8,26 +8,26 @@ const QuestionPresentation = ({ game, navigate, question }) => {
 
     useEffect(() => {
         const saveStats = async () => {
-        try {
-            const response = await fetch('http://localhost:8010/api/stats', {
-                method: 'POST',
-                headers: { 'Content-Type': 'application/json' },
-                body: JSON.stringify({
-                  correctAnswers: score.correct,
-                  incorrectAnswers: score.incorrect,
-                  totalRounds: maxRounds
-                })
-            });
-            
-            if (!response.ok) throw new Error('Error al guardar estadísticas');
-            
-        } catch (error) {
-            console.error('Error:', error);
-        }
+            try {
+                const response = await fetch('http://localhost:8001/api/stats', {
+                    method: 'POST',
+                    headers: { 'Content-Type': 'application/json' },
+                    body: JSON.stringify({
+                        correctAnswers: score.correct,
+                        incorrectAnswers: score.incorrect,
+                        totalRounds: maxRounds
+                    })
+                });
+    
+                if (!response.ok) throw new Error('Error al guardar estadísticas');
+            } catch (error) {
+                console.error('Error:', error);
+            }
         };
-
-        if (score.rounds >= maxRounds) saveStats();
+    
+        if (score.rounds === maxRounds) saveStats();
     }, [score]);
+    
 
     const checkAnswer = (selected) => {
         if (!question || buttonsDisabled) return;
